requirements.txt
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
